+++ Sources/VComponents/Components/Value Pickers/Slider/VSlider.swift
@@ -132,13 +132,16 @@ public struct VSlider: View {
     
     @ViewBuilder 
     private var thumbView: some View {
-        if uiModel.thumbDimension > 0 {
+        if 
+            uiModel.thumbSize.width > 0 &&
+            uiModel.thumbSize.height > 0
+        {
             Group(content: { // `Group` is used for giving multiple frames
                 ZStack(content: {
                     thumbBackgroundView
                     thumbBorderView
                 })
-                .frame(dimension: uiModel.thumbDimension)
+                .frame(size: uiModel.thumbSize)
                 .offset(
                     x: uiModel.direction.isHorizontal ? thumbOffset.withOppositeSign(uiModel.direction.isReversed) : 0,
                     y: uiModel.direction.isHorizontal ? 0 : thumbOffset.withOppositeSign(uiModel.direction.isReversed)
@@ -226,7 +229,9 @@ public struct VSlider: View {
     
     // MARK: Thumb Offset
     private var thumbOffset: CGFloat {
-        progressWidth - uiModel.thumbDimension/2
+        let thumbWidth: CGFloat = uiModel.thumbSize.dimension(isWidth: uiModel.direction.isHorizontal)
+
+        return progressWidth - thumbWidth/2
     }
 }